♻️ Refactor : RoomPost 수정
