+++ src/main/java/gaji/service/domain/Post/RoomBoard.java
@@ -1,4 +1,4 @@
-package gaji.service.domain.roomPost;
+package gaji.service.domain.Post;
 
 import gaji.service.domain.room.Room;
 import jakarta.persistence.*;
