+++ src/main/java/gaji/service/domain/Post/RoomComment.java
@@ -1,4 +1,4 @@
-package gaji.service.domain.roomPost;
+package gaji.service.domain.Post;
 
 import gaji.service.domain.enums.CommentStatus;
 import gaji.service.domain.studyMate.StudyMate;
